decoupling parsing function
