+++ BrokerageApi/V1/Gateways/ProviderGateway.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using BrokerageApi.V1.Gateways.Interfaces;
 using BrokerageApi.V1.Infrastructure;
-
+using BrokerageApi.Tests.V1.Gateways.Helpers;
 namespace BrokerageApi.V1.Gateways
 {
     public class ProviderGateway : IProviderGateway
@@ -22,7 +22,7 @@ namespace BrokerageApi.V1.Gateways
             return await _context.Providers
                 .Where(p => p.IsArchived == false)
                 .Where(p => p.Services.Any(s => s.Id == serviceId))
-                .Where(p => p.SearchVector.Matches(EF.Functions.ToTsQuery("simple", ParsedQuery(query))))
+                .Where(p => p.SearchVector.Matches(EF.Functions.ToTsQuery("simple", ParsingHelpers.ParsedQuery(query))))
                 .OrderBy(p => p.Name)
                 .ToListAsync();
         }
@@ -33,15 +33,5 @@ namespace BrokerageApi.V1.Gateways
                 .SingleOrDefaultAsync(p => p.Id == id);
         }
 
-        public static string ParsedQuery(string query)
-        {
-            var separators = new[] { " " };
-            var options = StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries;
-
-            var words = query.Split(separators, options).ToList();
-            var terms = words.ConvertAll(w => $"{w}:*");
-
-            return String.Join(" & ", terms);
-        }
     }
 }
